Refactored API. Added page selection. Split CSS into basic style and theme.
