+++ public/js/index.js
@@ -3,7 +3,6 @@
 var router = new VueRouter({
     routes: [
         { path: "/view/:id", component: Vue.component('main-view') },
-        { path: "/search/:tags", component: Vue.component('main-search') },
         { path: "/search", component: Vue.component('main-search') },
         { path: "/upload", component: Vue.component('main-upload') },
         { path: "/", redirect: "/search"}
