+++ src/js/api/delete-image-by-id.js
@@ -1,0 +1,7 @@
+const database = require("../database");
+
+module.exports = function(req, res) {
+    const imageID = req.params.id;
+    database.deletePicture(imageID);
+    res.sendStatus(200);
+};
