+++ src/js/api/delete-tag-from-image.js
@@ -1,0 +1,8 @@
+const database = require("../database");
+
+module.exports =  function(req, res) {
+    const imageID = req.params.id;
+    const tagName = req.params.tagname;
+    database.deleteTag(imageID, tagName);
+    res.sendStatus(200);
+};
