+++ src/js/api/get-tag-data-by-name.js
@@ -1,0 +1,10 @@
+const database = require("../database");
+
+module.exports = function(req, res) {
+    database.getCountByTagList(req.params.name.split(" "), function(data) {
+        res.send({
+            name: req.params.name,
+            count: data,
+        });
+    });
+};
