+++ src/js/api/post-new-image-by-url.js
@@ -1,0 +1,9 @@
+const database = require("../database");
+
+module.exports = function(req, res) {
+    database.insertPicture({
+        url: req.body.url,
+        tags: [],
+    });
+    res.sendStatus(200);
+};
