+++ src/js/api/post-new-tag-on-image.js
@@ -1,0 +1,8 @@
+const database = require("../database");
+
+module.exports = function(req, res) {
+    const imageID = req.params.id;
+    const tagName = req.params.tagname;
+    database.setTag(imageID, tagName);
+    res.sendStatus(200);
+};
