+++ src/js/api/post-new-tag.js
@@ -1,0 +1,6 @@
+const database = require("../database");
+
+module.exports = function(req, res) {
+    database.insertTag(req.params.name);
+    res.sendStatus(200);
+};
