+++ src/js/server.js
@@ -5,7 +5,6 @@ const fs = require('fs');
 const path = require("path");
 
 const confHelper = require("./config");
-const api = require("./api");
 const routes = require("./routes");
 const database = require("./database");
 
@@ -13,21 +12,14 @@ const app = express();
 app.use(bodyParser.json());
 app.use(busboy());
 
-try {
-    var configFilePath = path.resolve(__dirname, "../..", "config.json");
-    var configContents = fs.readFileSync(configFilePath, 'utf8');
-    console.log(`Found configuration file : ${configFilePath}`);
-    confHelper.initConfig(JSON.parse(configContents));
-} catch(e) {
-    confHelper.initConfig();
-}
+confHelper.initConfig();
 const config = confHelper.config;
 console.log(config);
 
 database.init();
 
 app.use("/img", express.static(config.imageFolder));
-api.init(app);
+app.use("/api", require("./api"));
 routes.init(app);
 
 app.listen(config.port, function() {
